Smart Resume Parser/requirements.txt
streamlit
spacy
pymupdf
python-docx
pandas
requests
matplotlib
